Update README.md
The protocol is effective in applications of BCI (Brain-Computer Interface), end-to-end verifiable voting systems, computing machine learning models via linear regression, and many other fields for providing full privacy to end-users while also uplifting the quality of human life. The area of machine learning on encrypted data is booming and is expected to grow significantly hence privacy is becoming a huge concern. 
AVISPA tool is used to analyze large-scale Internet security-sensitive protocols and applications for verifying and testing them. 
The protocol is tested for 2 to 5 clients and is scalable. To retain privacy and also mine data, the concept of partial homomorphic encryption is used.